+++ src/viewer/main/ViewerActionManager.C
@@ -92,7 +92,7 @@ public:
         // Add all of the action groups as choices.
         ViewerActionManager *actionMgr = window->GetActionManager();
         stringVector names(actionMgr->GetActionGroupNames(false));
-        for(int i = 0; i < names.size(); ++i)
+        for(size_t i = 0; i < names.size(); ++i)
             AddChoice(names[i].c_str());
     }
 
@@ -759,7 +759,7 @@ ViewerActionManager::EnableActions(ViewerWindowManagerAttributes *wma)
         // object and add the new action group to the manager.
         ActionGroup newAction(ag.GetName());
         newAction.enabled = ag.GetVisible();
-        for(int j = 0; j < ag.GetActions().size(); ++j)
+        for(size_t j = 0; j < ag.GetActions().size(); ++j)
         {
             ActionIndex index = ag.GetAction(j);
             if(index != ViewerRPC::MaxRPC)
@@ -852,7 +852,7 @@ ViewerActionManager::RealizeActionGroups(bool toolbarsVisible, bool largeIcons)
     for(;pos != actionGroups.end(); ++pos)
     {
         const ActionIndexVector &a = pos->actions;
-        for(int j = 0; j < a.size(); ++j)
+        for(size_t j = 0; j < a.size(); ++j)
         {
             ViewerActionBase *action = GetAction(a[j]);
             if(action && action->VisualEnabled())
@@ -939,7 +939,7 @@ ViewerActionManager::RealizeActionGroups(bool toolbarsVisible, bool largeIcons)
     //
     // Enable or disable the action groups based on user-settings.
     //
-    for(i = 0; i < actionGroups.size(); ++i)
+    for(i = 0; i < (int)actionGroups.size(); ++i)
     {
         SetActionGroupEnabled(i, actionGroups[i].enabled);
     }
@@ -992,7 +992,7 @@ ViewerActionManager::UpdateActionConstruction(ViewerActionBase *action)
     for(;pos != actionGroups.end(); ++pos)
     {
         const ActionIndexVector &a = pos->actions;
-        for(int j = 0; j < a.size(); ++j)
+        for(size_t j = 0; j < a.size(); ++j)
         {
             if(GetAction(a[j]) == action && action->VisualEnabled())
             {
@@ -1071,14 +1071,14 @@ ViewerActionManager::UpdateActionInformation(ViewerWindowManagerAttributes *wma)
 {
     wma->ClearActionConfigurations();
 
-    for(int i = 0; i < actionGroups.size(); ++i)
+    for(size_t i = 0; i < actionGroups.size(); ++i)
     {
         const ActionGroup &ag = actionGroups[i];
         if(ag.name != "Customize")
         {
             ActionGroupDescription description(ag.name);
             description.SetVisible(ag.enabled);
-            for(int j = 0; j < ag.actions.size(); ++j)
+            for(size_t j = 0; j < ag.actions.size(); ++j)
             {
                 ViewerActionBase *action = GetAction(ag.actions[j]);
                 if(action && !action->CanHaveOwnToolbar())
@@ -1198,7 +1198,7 @@ ViewerActionManager::UpdatePopup()
         const ActionIndexVector &a = pos->actions;
         bool hasEnabledActions = false;
         ViewerWindow *win = 0;
-        for(int i = 0; i < a.size() && !hasEnabledActions; ++i)
+        for(size_t i = 0; i < a.size() && !hasEnabledActions; ++i)
         {
             int actionIndex = a[i];
             ViewerActionBase *action = actions[actionIndex];
@@ -1296,11 +1296,11 @@ ViewerActionManager::AddAction(ViewerActionBase *action, ActionIndex index)
 void
 ViewerActionManager::AddActionGroup(const ViewerActionManager::ActionGroup &group)
 {
-    for(int i = 0; i < actionGroups.size(); ++i)
+    for(size_t i = 0; i < actionGroups.size(); ++i)
     {
         if(group.name == actionGroups[i].name)
         {
-            for(int j = 0; j < group.actions.size(); ++j)
+            for(size_t j = 0; j < group.actions.size(); ++j)
                 actionGroups[i].AddAction(group.actions[j]);
             return;
         }
@@ -1341,7 +1341,7 @@ ViewerActionManager::GetNumberOfActionGroupMemberships(ActionIndex index) const
         for(pos = actionGroups.begin(); pos != actionGroups.end(); ++pos)
         {
             const ActionIndexVector &a = pos->actions;
-            for(int i = 0; i < a.size(); ++i)
+            for(size_t i = 0; i < a.size(); ++i)
                 count += ((a[i] == index) ? 1 : 0);
         }
     }
@@ -1376,7 +1376,7 @@ void
 ViewerActionManager::SetActionGroupEnabled(int index, bool val, bool update)
 {
     // Look for the named action group.
-    if(index >= 0 && index < actionGroups.size())
+    if(index >= 0 && (size_t)index < actionGroups.size())
     {
         // Set the action group's enabled value.
         actionGroups[index].enabled = val;
@@ -1434,7 +1434,7 @@ ViewerActionManager::SetActionGroupEnabled(int index, bool val, bool update)
 bool
 ViewerActionManager::GetActionGroupEnabled(int index) const
 {
-    return (index >= 0 && index < actionGroups.size()) ? actionGroups[index].enabled : false;    
+    return (index >= 0 && (size_t)index < actionGroups.size()) ? actionGroups[index].enabled : false;    
 }
 
 // ****************************************************************************
@@ -1591,7 +1591,7 @@ void
 ViewerActionManager::ActionGroup::AddAction(ActionIndex index)
 {
     // Return early if the action is already in the list.
-    for(int i = 0; i < actions.size(); ++i)
+    for(size_t i = 0; i < actions.size(); ++i)
     {
         if(actions[i] == index)
              return;
