+++ src/viewer/main/ViewerServerManager.C
@@ -341,7 +341,7 @@ ViewerServerManager::AddArguments(RemoteProxyBase *component,
     //
     // Add arguments stored in ViewerServerManager.
     //
-    int i;
+    size_t i;
     for (i = 0; i < arguments.size(); ++i)
          component->AddArgument(arguments[i].c_str());
 
@@ -713,7 +713,7 @@ ViewerServerManager::OpenWithLauncher(
             // Search the args list and see if we've supplied the path to
             // the visit executable.
             std::string visitPath;
-            for(int i = 0; i < args.size(); ++i)
+            for(size_t i = 0; i < args.size(); ++i)
             {
                 if(args[i] == "-dir" && (i+1) < args.size())
                 {
@@ -859,7 +859,7 @@ ViewerServerManager::SimConnectThroughLauncher(const std::string &remoteHost,
             // Search the args list and see if we've supplied the path to
             // the visit executeable.
             std::string visitPath;
-            for(int i = 0; i < args.size(); ++i)
+            for(size_t i = 0; i < args.size(); ++i)
             {
                 if(args[i] == "-dir" && (i+1) < args.size())
                 {
